chore(fix): datepicker refactor
