+++ package.json
@@ -62,6 +62,7 @@
     "pod-install": "^0.1.7",
     "prettier": "^2.0.5",
     "react": "~16.9.0",
+    "react-dom": "^16.13.1",
     "react-native": "~0.62.2",
     "react-native-svg": "^12.1.0",
     "release-it": "^13.6.3",
@@ -167,7 +168,6 @@
     "dateformat": "^3.0.3",
     "expo-font": "^8.1.1",
     "expo-step-indicator": "0.1.2",
-    "react-dom": "^16.13.1",
     "react-native-safe-area-context": "~3.0.7",
     "react-native-slider": "^0.11.0",
     "react-native-typography": "^1.4.1"
