+++ src/components/DatePicker/DatePickerComponentType.ts
@@ -1,0 +1,7 @@
+export interface DatePickerComponentProps {
+  value: Date;
+  onChange: (e: any, data?: any) => void;
+  mode: "date" | "time" | "datetime";
+  toggleVisibility: () => void;
+  isVisible?: boolean;
+}
